Allow security-manager to build without SMACK support
When libsmack is not available on build platform,
some parts of security-manager dependent on it, won't be available.
In this case some functionality of security-manager related to SMACK
won't work.

It will be probably nice to introduce some kind of system rules layer
and allow using different types of mechanisms to secure the system.
One of such layers implementation could be based on SMACK.

Change-Id: Ie7c4b3982eaf9802e0754cc7dc6cf3e8843b9295